+++ .warp/setup/elasticsearch/elasticsearch.sh
@@ -20,6 +20,9 @@ then
         elasticsearch_version=$( warp_question_ask_default "Choose a version of elasticsearch: $(warp_message_info [5.6.8]) " "5.6.8" )
     
         case $elasticsearch_version in
+        '6.5.4')
+            break
+        ;;
         '6.4.2')
             break
         ;;
@@ -36,7 +39,7 @@ then
             break
         ;;
         *)
-            warp_message_info2 "Selected: $elasticsearch_version, the available versions are: 6.4.2, 5.6.8, 2.4.6, 2.4.4, 1.7.6"
+            warp_message_info2 "Selected: $elasticsearch_version, the available versions are: 6.5.4, 6.4.2, 5.6.8, 2.4.6, 2.4.4, 1.7.6"
         ;;
         esac        
     done